GLK: Fix compiler warnings

+++ engines/glk/frotz/detection.cpp
@@ -99,7 +99,7 @@ bool FrotzMetaEngine::detectGames(const Common::FSList &fslist, DetectedGames &g
 					fname = Common::String(fname.c_str(), dot);
 
 				debug("ENTRY0(\"%s\", \"%s-%s\", \"%s\", %u),",
-					folderName.c_str(), fname.c_str(), serial, md5.c_str(), filesize);
+					folderName.c_str(), fname.c_str(), serial, md5.c_str(), (uint)filesize);
 			}
 			const PlainGameDescriptor &desc = FROTZ_GAME_LIST[0];
 			gd = DetectedGame(desc.gameId, desc.description, Common::UNK_LANG, Common::kPlatformUnknown);
